improve-function
If secrets key already exist, we dont call S3

+++ amplify/backend/function/amplifygcpwebrisk/src/index.js
@@ -4,11 +4,6 @@
 exports.handler = async (event) => {
     console.log(`EVENT: ${JSON.stringify(event)}`);
 
-
-    const AWS = require('aws-sdk');
-    const S3 = new AWS.S3();
-    const fsPromise = require('fs').promises;
-
     let bucket = process.env.BUCKET
     console.log("bucket: " + bucket);
     let gcpKey = process.env.GCP_WEB_RISK_KEY
@@ -16,26 +11,42 @@ exports.handler = async (event) => {
     let destFile = process.env.GOOGLE_APPLICATION_CREDENTIALS
     console.log("destFile: " + destFile);
 
-    var credentials = ""
-    try {
-        const params = {
-            Bucket: bucket,
-            Key: gcpKey,
-        };
+    // First check if the creds already exist locally
+    const Fs = require('fs');
+    let credsExist = Fs.existsSync(destFile);
 
-        credentials = await S3.getObject(params).promise();
-        await fsPromise.writeFile(destFile, credentials.Body);
-        console.log("credentials: " + credentials);
-        console.log("GCP secrets loaded");
+    console.log("GCP secrets  exist?: " + credsExist);
 
-    } catch (err) {
-        console.log("Error retrieving GCP credentials: " + err);
+    // If creds does not exist, lest retrieve it from S3
+    if (!credsExist) {
+
+        try {
+            const AWS = require('aws-sdk');
+            const S3 = new AWS.S3();
+            const fsPromise = require('fs').promises;
+
+
+
+            const params = {
+                Bucket: bucket,
+                Key: gcpKey,
+            };
+
+            credentials = await S3.getObject(params).promise();
+            await fsPromise.writeFile(destFile, credentials.Body);
+            console.log("credentials: " + credentials);
+            console.log("GCP secrets loaded");
+            credsExist = true;
+        } catch (err) {
+            console.log("Error retrieving GCP credentials: " + err);
+        }
     }
 
     let statusCode = 500;
     let body = "";
 
-    if (credentials !== "") {
+    // Once we have the creds we can do the call to GCP Web Risk API
+    if (credsExist) {
 
         const { WebRiskServiceClient, protos } = require('@google-cloud/web-risk');
         const client = new WebRiskServiceClient()
@@ -73,6 +84,7 @@ exports.handler = async (event) => {
         body = "There was an error retrieving the S3 GCP credentials file"
     }
 
+    // Finally we return
     return {
         statusCode: statusCode,
         body: body,
